requirements.txt
atari-py==0.2.9
gymnasium[atari]==0.29.1
kornia==0.7.1
urllib3==1.26.11
wandb
numpy==1.24.4
Pillow==10.2.0
matplotlib==3.7.5
rl-zoo3==2.2.1
scikit-learn==1.3.2
seaborn==0.13.2
stable-baselines3[extra]==2.2.1
torch==1.13.1+cu117
torchvision==0.14.1+cu117
tensorboard
tqdm==4.66.2
pyyaml
ray[tune]
avalanche-rl @ git+https://github.com/ContinualAI/avalanche-rl.git
moviepy==1.0.3